[TOOL-114] Split up parser.ts into multiple files
This is the very first step of cleaning up the parser.
