+++ lib/src/generators/contract/json-schema.spec.ts
@@ -19,7 +19,7 @@ import {
   unionType,
   VOID
 } from "../../models";
-import { parsePath } from "../../parser";
+import { parsePath } from "../../parsing/file-parser";
 import { generateJsonSchema, jsonTypeSchema } from "./json-schema";
 
 const EXAMPLES_DIR = path.join(
