+++ lib/src/generators/contract/openapi3.spec.ts
@@ -1,6 +1,6 @@
 import * as fs from "fs-extra";
 import * as path from "path";
-import { parsePath } from "../../parser";
+import { parsePath } from "../../parsing/file-parser";
 import { generateOpenApiV3 } from "./openapi3";
 
 const EXAMPLES_DIR = path.join(
